# lib/api-zod/src/generated/types/setSong.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SongBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SetSong {
  id: string;
  title: string;
  artist: string;
  /** @nullable */
  meta?: string | null;
  sortOrder: number;
}