# lib/api-zod/src/generated/types/songUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SongBook API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SongUpdate {
  title?: string;
  artist?: string;
  meta?: string;
  lyricsChords?: string;
  /** @nullable */
  spotifyLink?: string | null;
}